Re-define the meaning of cache_size to avoid expensive and complex updates.
Now we redefine cache_size to not mean 1+ offset of the current last
uptodate byte but to 1+ offset of the last uptodate byte this cache has
seen. cache_size is increased on set_uptodate() but it's only decreased
in truncate() and not everytime we release()d some chunk from the cache.
This simplifies updation of cache_size.